Enable the TEST option by default in atf-based ports.
Get ports that use the atf libraries to install their tests by default.
This is to showcase the potential of getting the tests installed on a
live system and will also simplify the process of running them periodically
out of /usr/local/tests/.

Packages confirmed to build with this option enabled.

Approved by:	bapt (ports)