comms/dcf77pi: Update 3.0.0.1
The code to decode the radio signal, date/time, and civil warnings is now
provided in a library libdcf77.so together with the relevant header files.

dcf77pi is now split up into dcf77pi (the live client) and dcf77pi-analyze
(the client to analyze log files). The -f parameter of dcf77pi is dropped.

Apply minor lint(1) fixes.