Properly respect CFLAGS, LDFLAGS, etc.
When building a port with base system make, variables like CC and CFLAGS
need to be defined using '=' in the makefile.  With '+=' values from
make.conf can be picked up.  Normally the ports system values for these
variables are passed to a configure script which generates a makefile with
them.  If there's no configure script the makefile has to be patched to use
the right values.