Add explicit dependency on qt4-gui.
This is similar to r356888, in that PyQt's configure.py builds py-qt4-test
with QtGui's libraries (even though, contrary to qt4-sql, it is not used at
configuration time). Using QtGui defines QT_GUI_LIB, which makes the QtTest
headers pull in QtGui ones. QtGui comes for free indirectly when the API
option is set, but the build fails otherwise:

  c++ -c -O2 -pipe -fno-strict-aliasing -fPIC -O2 -Wall -W -pthread -D_THREAD_SAFE -DNDEBUG -DQT_NO_DEBUG -DQT_TEST_LIB -DQT_GUI_LIB -DQT_CORE_LIB -I. -I/usr/local/include/python2.7 -I/usr/local/share/qt4/mkspecs/default -I/usr/local/include/qt4/QtTest -I/usr/local/include/qt4/QtGui -I/usr/local/include/qt4/QtCore -I/usr/local/include/qt4 -I/usr/local/include -o sipQtTestQTest.o sipQtTestQTest.cpp
  In file included from /wrkdirs/usr/ports/devel/py-qt4-test/work/PyQt-x11-gpl-4.10.3/sip/QtTest/qtestkeyboard.sip:31:
  /usr/local/include/qt4/QtTest/qtestkeyboard.h:56:10: fatal error: 'QtGui/qapplication.h' file not found

PR:		183706
MFH:		2014Q2