Resurrect devel/c2mdoc (with stage support), assign maintainer
It seems that this port was killed for not having publically available
distfiles, but from what I can tell, this was always a distfile-free
port, meaning that it was removed in error.

The PR to bring it back did not provide the required stage support, but
luckily that was trivial to add.  Differences from the previous version:

  * pkg-descr rewrapped
  * whitespace removed from awk script (caused PORTREVISION bump)
  * Header trimmed
  * NOMAN definition removed
  * CP+REINPLACE_CMD replaced with SED and moved to post-extract target
  * Stage support
  * Maintenance passed to PR submitter

PR:		189009
Submitted by:	Chris Hutchinson
Tweaks by:	marino