games/linux-steam: Reset maintainer and deprecate this interactive port
The maintainer is not reachable via email although messages don't bounce.
The current and previous PR have both timed out.  Reset the maintainer and
set deprecation date for 20 August, the same the other unmaintained
interactive ports.  The port is already staged, but a new maintainer
should convert this interactive port to a regular one in order to
undeprecate it.

PR: 191219