devel/scala-ide: Update distinfo
The last update never fetched properly as the tarball was likely rerolled.
I visited the master site and drilled down to the official download
directory where a sha1 hash was provided for the version 3.0.4 download.
I verified the downloaded file had the same sha1 digest listed in the
directory (which actually doesn't mean much because anyone capable of
compromising a tarball on the official download site can also compromise
the digest, but it's the best I can do without the first version.)

No revbump because fetch never passed for anyone.