Add patches to fix the build with base's ancient GCC.
In practice, this fixes the build on 8 and 9 when clang is not used. There
is no need to submit the patches upstream because the next version of Amarok
will already require a C++11 compiler, which makes the problems disappear.

I didn't feel the need to bump PORTREVISION because the generated code
should be the same, and users who had already managed to build Amarok don't
need to rebuild it.

PR:		192159
Submitted by:	freebsd@chillt.de