Fix transmission-cli,-gtk on some platforms after openssl update
At least on some releases, transmission-cli and transmission-gtk now
need -lcypto in LDFLAGS to build.  Additionally transmission-gtk needed
an update of the inotify configure argument.  The latter requires a
revbump because it seems possible that it could have been built with
the bad argument somewhere.

PR:		192319
Submitted by:	Ivan Klymenko
Approved by:	"Just Fix It" blanket, but maintainer was in copy on PR