cad/cura-engine: Engine of slicing solution for RepRap 3D printers
The CuraEngine is a C++ console application for 3D printing GCode generation.
It has been made as better and faster alternative to the old Skeinforge engine.

The CuraEngine is pure C++ and uses Clipper from
http://www.angusj.com/delphi/clipper.php. There are no external dependences
and Clipper is included in the source code without modifications.

This is just a console application for GCode generation. For a full graphical
application look at https://github.com/daid/Cura with is the graphical
frontend for CuraEngine.

The CuraEngine can be used seperately or in other applications.
Feel free to add it to your application. But to take note of the License.

WWW: http://wiki.ultimaker.com/Cura

PR:		192486
Submitted by:	cederom tlen pl