databases/freetds: Fix build with libiodbc, reset maintainer
The maintainer (cowbert) hasn't been seen since 2009 and several PRs
timed out in the span of months.  Reset freetds and it's slave port.

Moveover, it's ancient.  All the MASTER_SITES are gone.  I had to
switch this over to Fedora repo so that it had public distfiles again.

The actual PR fixes the build when libiodbc is selected.
There is a freetds-devel port that's much newer.  I recommend that it
move these location and they disappear.

PR:		192224
Submitted by:	Denys Razumov
Approved by:	maintainer timeout (of course)