On FreeBSD 9.x i386 systems, debugging applications with threads causes a crash of gdb.
When the libthr is loaded by gdb, the thread support want to add FreeBSD
specific information to the current thread. Doing this, it cleans the
thread list and re-add the current thread with the updated information
(actually, the LWP id).
This operation call a free of the struct thread_info (clean/init the
thread list) and a malloc of the same struct (add the current thread).
The problem is that a pointer to the struct thread_info is stored in
another struct (about the execution context) and it's not updatable.

On FreeBSD 9.x i386 systems, the malloc return a new pointer; then the
execution context contains a broken pointer.
On all other platform, the malloc return the same pointer and the
execution context stays coherent.

This patch avoids to clean the thread list and to re-add the thread; it
just update thread information of the current thread.

PR:		192890
Submitted by:	luca.pizzamiglio@gmail.com (maintainer)