devel/ocfpcsc: Expand license permissions
The license added to the PR was unnecessarily restrictive and caused
the builders to no longer package the binaries.  I looked at the IBM
licence provided in files and did not see any restriction on packaging
object files.  I alerted the maintainer about this twice, and wg@ told
me to Just Fix It.

PR:		191419
Approved by:	maintainer timeout (Just Fix It)