Older make (8 and 9) treats a string as one large string. Newer make (10+) treats a string as a collection of words.
Somehow a double-space was introduced into CFLAGS, which caused gdb66's
configure to choke. Newer make had to be told to examine the entire line
to get rid of it, but that syntax (/W) was unknown to older make.

So, split the difference and just make a shell call to remove the space.
The port can now build on 10 and 11.