security/hydra: Switch Firebird client options from 2.1 => 2.5
databases/firebird21-* will be removed imminently.  Hydra has a
non-default option that relies on the firebird client library which
the currently version of Firebird (2.5.x) also has.  This change
is untested.  There's no revbump because the option is not normally set.

Change is necessary, it's an attempt to avoid breakage when Firebird
ports are removed.