www/mohawk: Unbreak on DragonFly, remove useless LDFLAG
The provided LDFLAG of -L${LOCALBASE}/lib/event2 would have stopped
working after libevent2.so moved to ${LOCALBASE}/lib had the LDFLAG
affected anything.  The LDFLAG is set by Makefile.inc, and the
handling of DragonFly is wrong for ports, although it might still be
correct for pkgsrc.  Fix that with a patch and remove the LDFLAG
that doesn't do anything.  No revbump required.