security/wpa_supplicant: Fix NDIS driver, privsep no longer default
The NDIS driver didn't build when privsep option was turned off.  Moveover,
it didn't display the driver name correctly, so I figured out a patch
based on base's ndis driver code.  Thirdly, when privsep option is turned
off, wpa_priv is not built, so let's handle that condition.

Finally, remove privsep from default options because it removes the
standard list of drivers and replaces it with privsep driver.