wpa/supplicant: Fix driver list with NDIS, take 2
It turns out the driver list was only showing "null" for NDIS when -h
(help) or an unknown switch was passed to wpa_supplicant.  The cause is
that the NDIS driver is the only that that has to be initialized, and
the initialization wasn't occurring in the above case.

The fix is the move the NDIS driver initialization before the command
switches are examined in case wpa_supplicant aborts to usage early.