japanese/ebnetd: Remove @dirrm, use @sample, use @dir()
The stage-qa checks said 5 files needed perl shebang fixed, but
USE=shebangfix won't fix it.  The value of PERL was defined for the
configure target to fix it.  However, a run depends on perl was not
added as the affected files are probably optional scripts.

PR:		194367
Submitted by:	maintainer (Yasuhiro KIMURA)