finance/libstocks: Remove --mmap switch from grep command
This is a GNU grep specific switch that was disabled in March 2010
(~version 2.6) and finally removed in version 2.17 (Feb 2014).  It's
still supported in the base grep, but the man page talks about possible
core dumps.  DragonFly has grep v2.20 in base now and the port stopped
building.  Just remove the deprecated/EOL'd option.