Remove dns/powerdns-devel port per maintainer approval
As dns/powerdns was just updated to version 3.4, there was a decision to
be had to disable the -devel version or just remove the port altogether.
Due to the frequency of releases and the manner on how the upcoming
version is tested, it made sense to retire the port (at version 3.3).

PR:		194508
Submitted by:	maintainer (Ralf van der Enden)