Remove ZEROCONF option
mDNSResponder was abbandoned by upstream and replaced by discoveryd, which is
not open source. As kdelibs already depends on AVAHI being installed, remove
the option for the zeroconf backend and always use AVAHI.

Also, port option processing is broken, and AVAHI is used even if
mDNSResponder is chosen as backend, so bump PORTREVISION as the old package
invalidly tracks an unreferenced mDNSResponder library package dependency.

Approved by:	rakuco (mentor)