Add new ports games/openclock
PR:		194617
Submitted by:	Kevin Zheng

OpenClonk is a free multiplayer action game in which you control clonks,
small but witty and nimble humanoid beings. The game is mainly about
mining, settling and fast-paced melees.

OpenClonk is not just a game but also a versatile 2D game engine that
allows the creation of mods. It is the successor of the shareware game
series Clonk and thus inherits many of its features.