Fix the ancient problem, where we replaced Linux' "cp -pd" with simple "cp -p". This caused symbolic links (like libFOO.so -> libFOO.so.X.Y) to be dereferenced installing three copies of each library instead of one.
Strip the executables and the shared libraries after staging.

Fix the authors' makefiles to allow parallel building.

Add a patch which allows us to replace the multitude of
flite_VOICE-executables (one for each voice) with sym-links to the
single flite itself.

The size of the package shrunk from 100Mb to 50Mb here (amd64) as the
result of the above measures.

Bump PORTREVISION.

Note: the new patches are deliberately organized by feature rather than
by the patched files -- to make it easier to refer the upstream
maintainers to them.