Update "BSD" license
The license included by the author is the CPL, but not the "Common
Public License" as indicated in the port Makefile. Instead, it is the
Cypherpunk Anti-License which is about as much of a non-license license
you can get. The terms indicate you can redistribute/relicense the code
in any license you want, but the author for some strange reason gives
his opinion on which license he would recommend. This makes no sense
given the spirit of the CPL.

Kids, don't do drugs.