Remove port databases/postgresql93-plv8js (avoids broken index)
This port was added on 5 October 2014.  The intention for its existence
was to provide a way to use plv8js with pgsql 9.3 instead of the default
pgsql 9.2.  It was implemented in such a way that if PGSQL_DEFAULT is
set to 9.3, the index breaks with a duplicate origin with datbases/
postgresql-plv8js.  It's possible to adjust the plv8js ports by converting
the version into an option and using typical master/slave techniques, but
I can't come up with a good reason to do this at all.

I don't think this port ever should have been created.  Anyone that would
need this port would have needed to set PGSQL_DEFAULT anyway (which
already works).  In the worst case, WANT_PGSQL could be based through a
command line.  Perhaps the motivation was to have a binary package to
avoid building it, but this reason disappears soon when the default
version of pgsql is bumped to 9.3.  Based on all those reasons, I think
it is better to remove the port outright (pointing to master port) rather
than adjust it to avoid a broken index.

PR:	195281