New port comms/xcwcp Sub-port of comms/unixcw
Previously, this was a build-time option, but the primary use of
comms/unixcw is to provide libcw.so (and it's a dependency of
comms/cwdaemon).

Since comms/cwdaemon is not a GUI program, it doesn't make sense to enable
the GUI by default, which would mean pkg users couldn't install this
program.  By splitting it into a sub-port, it's now available via pkg
and more obviously present when earching ports.