multimedia/x264: Use libx264's config headers
An encoding error condition is created when HI10P is enabled
in libx264, since the x264 is not built with that options.

Make x264 use the headers provided by the libx264 port, not the ones
that come with the sources, so the configuration options between the
two ports (libx264 and x264) can't end up mismatched.

Thanks to astrodog and a_b (IRC) for coming up with an elegant solution. [1]

While I'm here:

- Update COMMENT to reflect this is just the x264 CLI frontend
- Add LICENSE_FILE
- Tweak BUILD_DEPENDS
- Sort USES and OPTIONS
- Update GCC_DESC to match libx264
- Use OPTIONS helpers (except for PGO & GCC .. *shakes fist*)
- Sort PLIST_*
- Remove conditional LDFLAGS (-fstack-protector is now handled by the
  framework)
- Whitespace alignment
- Remove pkgconfig REINPLACE_CMD, its not needed here.
- Strip x264 binary x264

PR:		195292 [1]
Submitted by:	Harrison Grundy <harrison dot grundy at astrodoggroup com>