Stop installing an empty headers directory.
A fix has been submitted to https://github.com/FreeRDP/Remmina/pull/428.

While a proper decision is not made there, just patch the code to stop
calling INSTALL(DIRECTORY ...) with a path that does not exist. Until CMake
3.1.0, this would end up creating an empty directory (hence the @dir entry
in the plist). The behavior has changed in 3.1.0, and we were getting build
errors because we were referring to a directory that did not exist in the
plist.