net/kamailio: install two files necessary if build with PQSQL
If built with the PGSQL option, two files necessary for creating
the initial database are not installed as necessary. This causes
"kamdbctl create" to fail during extra_create() when EXTRA_MODULES
contains "mohqueue" and/or "rtpproxy" (which is does, by default).

PR:		196446
Submitted by:	John W. O'Brien <john@saltant.com>