- Add p5-mem 0.4.5
mem is a trivial pragma to either allow defining the module it is included from
as being defined so that later classes or packages in the same file can use the
package to pull in a reference to it, or to be able to call its import routine
from a different package in the same file.

With parameter assignments or other actions, it forces those assignments to be
done, immediately, at compile time instead of later at run time. It can be use,
for example, with Exporter, to export typed-sub's among other usages.

WWW: http://search.cpan.org/dist/mem/