databases/p5-DBIx-Class-Helpers: UNBREAK
I'm not sure what happened here.  There was a good PR [1] to update
this port that properly changed the Makefile, distinfo, and the pkg-plist,
but tj@ updated the port independently but only changed the Makefile
(and moved PORTREVISION to 2 instead of resetting it to 0).  Use the
changes outlined in the PR properly updates the version to 2.024001.

PR:		195195
Submitted by:	Geoffroy Desvernay
Approved by:	blanket (unbreak port)