mail/pine-pgp-filters: use gnupg only, fix MASTER_SITES
When dougb released the port, he also reset the MASTER_SITES.  It has
been pulling off the cache ever since.  If the cache gets cleared, this
port would break.  I've moved the distfiles to my LOCAL site so this
never happens.

The other big change is to remove this hacky dependency logic.  Someone
tried to select between gnupg and gnupg1 depending on what was already
installed.  For the standard packages builders, this meant the port always
used the old gnupg1 because gnupg was never installed in the clean jail.

As both gnupg packages are range between 1.0M and 1.5M in size and can
coexist, the savings of resources in miniscule assuming they weren't both
installed anyway.  Let's just pick the new gnupg as the unconditional
dependency and remove the hack.  Inspired by discussion on PR which
was trying to fix a bug in the hack.

PR:		195426
Submitted by:	Trond Endrestol