[NEW] www/py-django-contrib-comments: Code formerly known as django.contrib.comments
Django used to include a comments framework; since Django 1.6 it's been
separated to a separate project. This is that project.

This framework can be used to attach comments to any model, so you can use
it for comments on blog entries, photos, book chapters, or anything else.

WWW: https://github.com/django/django-contrib-comments