devel/ptlib: Fix support for gcc-5
The ptlib port has a bug in a macro condition.  The condition wants
to include a different c++ header if GCC 4.1 or below is used.  Since
this compiler doesn't exist in ports and is ancient, I'm just going
to remove the check rather than fix it.  This change fixes the build
for the upcoming gcc-5 release.

Approved by:	just-fix-it blanket