deskutils/treesheets: make environment was not set in build target
This port's makefile fails to set the make environment when it overrode
the build target.  Given that the 2-line build.sh script has to be
patched in the post-configure target, the whole script should move to
the port makefile.  However, for now just set the make environment which
fixes the build on DragonFly with gcc-5.

Approved by:	just-fix-it blanket