Add www/p5-CGI-Fast.
This is the continuance of the CGI::Fast module from the CGI distribution.

CGI::Fast is a subclass of the CGI object created by CGI.pm. It is specialized
to work with the FCGI module, which greatly speeds up CGI scripts by turning
them into persistently running server processes. Scripts that perform
time-consuming initialization processes, such as loading large modules or
opening persistent database connections, will see large performance
improvements.

If there are no problems with this port, in a week I'll transfer maintainership
to perl@.