multimedia/mkvtoolnix: set environment during build/install and unsuppress
This port was strange in that it not only failed to set the environment,
it was unsetting PREFIX to whatever environment it had.  This change sets
the environment in both build and install targets with the exception of
PREFIX.  I don't know why that is done, but I maintained that part.

While here, unsuppress the commands so that the build can be logged and
change ";" to "&&" conjunctions which is preferred.

Approved by:	blanket