Mk/Uses/ada.mk: Add gcc5-based Ada compiler support
A new argument has been added to USES=ada, "5", which will select the
gcc5-aux compiler (gcc 5) instead of the gcc-aux compiler (gcc 4.9).

Additionally, it will check if ADA_DEFAULT is defined and equal to "5".
If that is the case, gcc5-aux will be used by default, even if no
arguments are defined.  The use case is to define it in make.conf so that
all Ada ports are built with same compiler.  This capability will be used
to confirm all Ada ports build correctly with gcc5-aux before the default
compiler is switched.