lang/gcc5-aux: Disable diagnostic checks
These diagnostic checks cause lang/gcc-aux to take nearly two hours
building as with 3 stages.  They are pretty heavy; I'm going to disable
them (with the default checking for releases).  FWIW, lang/gcc5 still
has these checks.

Technically this requires a revbump, but given the compiler is experiment
and that it takes two hours to build, I'll make the decision not to bump.