Upgrade to apache-openoffice SVN revision 1667428.
Word around broken build on FreeBSD 11.0-CURRENT i386, which appears
to be a code generation bug in clang 3.6 when using -Os optimization.

Build using the ports version of graphics/silgraphite instead of
building an internal copy.

When building with clang, use the ports version of devel/boost-libs
instead of an internal copy.

Use the dictionaries already installed by ports.

Update the patch-bridges and patch-*.mk patches to versions that
will be sent upstream.  The latter are tuned for clang to use as
is, with only some minor additional tweaks needed for gcc.

Nuke patch-webdav because it has been incorporated upstream.

Wordsmith pkg-message.

Various Makefile cleanups.

Differential Revision:	https://reviews.freebsd.org/D2108
Reviewed by:	pfg
Approved by:	mat (mentor)