Add SMHasher hashing benchmark and test system as a new port.
SMHasher is a test suite designed to test the distribution, collision,
and performance properties of non-cryptographic hash functions -
it aims to be the "DieHarder" of hash testing, and does a pretty
good job of finding flaws with a number of popular hashes.

The SMHasher suite also includes MurmurHash3, which is the latest
version in the series of MurmurHash functions - the new version is
faster, more robust, and its variants can produce 32- and 128-bit
hash values efficiently on both x86 and x64 platforms.

https://code.google.com/p/smhasher/