Rename multimedia/mediabrowser to multimedia/emby-server, and update to 3.0.5572.0.
The user and group names have change, but the uid/gid remain the same.
The port will continue to use the old database and user/group names if
they exist, but will default to the new names for new installs.

Instructions are in pkg-message for how to manually migrate things
yourself.

PR:		199093
Submitted by:	maintainer (Ben Woods)