audio/sooperlooper: Unbreak on some platforms
This unmaintained port is breaking on Wx 3.0, which is only an GUI option.
Downgrading to Wx 2.8 allows it to build on all platforms (as well as
having the option by default).  The second iteration of the PR suggested
Wx3.0 for FreeBSD 8 and 9, but Wx  2.8 for F10+, but I didn't like that
approach.  It only works on F8 and F9 because of compiler deficiencies.
Simple attempts of setting -std to c++11 didn't work, of course.

PR:		199221
Submitted by:	ports fury