lang/gcc5-aux: Build in one stage by default, new bootstrap for DF
The original intention was to build update all the bootstrap compilers
to be based on GCC5, but right now I'm only doing that for DragonFly.
After June 30, I might rebuild the bootstrap compiler on FreeBSD 10.1,
but right now gcc5-aux is building fine on FreeBSD in one stage, even
though their bootstrap is based on GCC 4.9.

Right now, gcc5-aux is building with 3 full stages.  I'm switching it to
a single stage build, but adding an off-by-default option to build it
in three stages again.  With one stage, libcc1 is not built, but with
three stages it is.  LibCC1 is meant to support the upcoming release of
GDB so it's currently not needed (it fails to build if gcc is not fully
bootstrapping which is why it's disabled)

This brings the build down to about 9-14 minutes versus 60-90 minutes
depending on the (average power) machine.

Also, I've explicitly limited the build to i386 and amd64 on FreeBSD and
I have dropped i386 support on DragonFly, which could only have been
build on DragonFly 3.8 or earlier.

Also, I removed the iconv dependency when a bootstrap compiler is being
created (the all-overriding BOOTSTRAP option for the maintainer)