Quisk Version 3.6.21 March 2015 =============================== Quisk can now transmit a message from a WAV file.  Record your message at a high level (near clipping) at 48 ksps, 16-bit, one channel (monophonic). Then enter the file name on the Config/Config screen.  Press the "File play" button to transmit.  Quisk will press the PTT button for you, and release it during pauses.  To interrupt playback, press PTT or release FilePlay so you can answer.
The "Split" button has been replaced with a "Splt" button and a "Rev" button.  The "Splt"
button splits Rx and Tx; and if you click it with the right mouse button instead of the
left, it also locks the Tx frequency so tuning changes the Rx frequency.  The "Rev"
button reverses the Tx and Rx frequencies.  These features were suggested by Mario, DH5YM.

I added a new parameter mic_agc_level to the config file to control the mic AGC.  Input levels
below mic_agc_level are ignored.  The default is 0.1.  Increase this (up to 1.0) to reduce the
AGC range, and reduce it to increase the AGC mic gain boost.

Philip Lee contributed a patch to the PulseAudio code to set the play buffer size.

Prompted by:		portscout